api_gateway/requirements.txt
uvicorn == 0.32.0
fastapi == 0.115.5
pika == 1.3.2
python-dotenv == 1.0.1